front-dashboard/requirements.txt
altair
pandas
streamlit
numpy
pymongo
python-dotenv
jsonpath-ng